requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
requests==2.22.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-dotenv==0.20.0
